コンポーネントの親子関係を実態に沿う形に変更
App 配下に List と EditMemoPanel を siblings として配置した
それに伴い全ての state をリフトアップした
