+++ src/EditMemoPanel.js
@@ -1,14 +1,53 @@
 export default function EditMemoPanel({
   status,
-  memoContent,
-  submittingEmptyMemo,
-  onMemoTextAreaChange,
-  handleEditButtonClick,
-  handleDeleteButtonClick,
+  updateStatus,
+  memos,
+  updateMemos,
+  memo,
+  updateMemo,
 }) {
+  function clearMemoState() {
+    updateMemo({ id: null, content: "" });
+  }
+
+  function handleEditButtonClick() {
+    if (!memo.content) {
+      return;
+    }
+
+    const id = memo.id;
+    const { content } = memo;
+    const title = content.split(/\r\n|\n|\r/)[0];
+    localStorage.setItem(id, content);
+
+    const newMemos = memos.map((m) => {
+      if (m.id === id) {
+        return {
+          id,
+          title,
+          content,
+        };
+      }
+      return m;
+    });
+    updateMemos(newMemos);
+    updateStatus("viewing");
+    clearMemoState();
+  }
+
+  function handleDeleteButtonClick() {
+    localStorage.removeItem(memo.id);
+    updateMemos(memos.filter((m) => m.id !== memo.id));
+    updateStatus("viewing");
+  }
+
+  function onMemoTextAreaChange(e) {
+    updateMemo({ id: memo.id || null, content: e.target.value });
+  }
+
   return (
     <div className="section">
-      <textarea value={memoContent || ""} onChange={onMemoTextAreaChange} />
+      <textarea value={memo.content || ""} onChange={onMemoTextAreaChange} />
       <button
         className="add-button"
         type="submit"
@@ -25,7 +64,7 @@ export default function EditMemoPanel({
           削除
         </button>
       )}
-      {submittingEmptyMemo ? (
+      {memo.content === "" ? (
         <div className="error-message">空のメモは保存できません</div>
       ) : null}
     </div>
